+++ services/admin/src/routes/index.svelte
@@ -70,7 +70,7 @@
 
   <br />
   <br />
-  <a href="/object"><button class="primary">Create New Object</button></a>
+  <a href="/object/new"><button class="primary">Create New Object</button></a>
 
   <br />
   <br />
